Update plugins/kaniko and node versions in drone.yml
This commit updates the version of the 'plugins/kaniko' and the 'node' images in the .drone.yml configuration file. The 'plugins/kaniko' is updated to 1.8.8-kaniko1.9.1 and 'node' to 20-alpine. These updates apply to all the relevant pipeline steps in the configuration, enhancing the overall build process with newer image versions.