Update Docker and launch settings configurations
Paths in Docker's watch function have been adjusted to target specific directories accurately. Additionally, an empty `RedisHost` attribute has been added in the Api's launch settings. This ensures proper tracking of changes in the FrontEnd and Hive directories, and prepares for the inclusion of Redis as a service in the application.